requirements.txt
python-telegram-bot==20.3
mediapipe
opencv-python
Pillow
numpy
